shorter timeout
Longer timeouts made it really annoying to wait for test results
when there were any failures with Compose. So the timeout is
shorter, but we'll run docker-compose alone before the tests
to make sure that everything's pulled and that there are no
problems.

Pytest-docker will require more robustness and more view of
what's currently happening in Compose, so we won't have to dance
around its limitations.